# [SPARK-24945][SQL] Switching to uniVocity 2.7.2