--- a/sql/core/pom.xml
+++ b/sql/core/pom.xml
@@ -38,7 +38,7 @@
     <dependency>
       <groupId>com.univocity</groupId>
       <artifactId>univocity-parsers</artifactId>
-      <version>2.6.3</version>
+      <version>2.7.2</version>
       <type>jar</type>
     </dependency>
     <dependency>