@MaxGekk 2018-07-27
@HyukjinKwon @maropu Please, take a look at the PR. Is it valid to just count number of lines returned by Hadoop LineReader and do not call parser at all? Maybe there are some corner cases when parser must be called?