@gatorsmile 2018-07-30
@MaxGekk @HyukjinKwon We are unable to merge this PR since the performance regression is very obvious.