@HyukjinKwon 2018-07-30
Ah, that looks going to be addressed in https://github.com/apache/spark/pull/21909 if you refer the number for `count()` which's not going to execute the parsing path that this upgrade affects.