@gatorsmile 2018-07-30
@HyukjinKwon We need to rerun the perf tests after https://github.com/apache/spark/pull/21909 is merged. 

We are also unable to accept the perf regression larger than `5%`. Based on the description, we still face a performance regression besides `count()`