@gatorsmile 2018-07-30
@HyukjinKwon I would suggest to skip this upgrade and then we can get 3.5 times perf improvement for `count()` .