@HyukjinKwon 2018-07-30
> we can get 3.5 times perf improvement for `count()`.

We shouldn't merge this as is for clarification. it needs a close look after #21909 which avoids the parsing code path by Univocity. How about this: after #21909, we should rerun the benchmark and see the numbers, and decide if we should reject this one or not.