@jbax 2018-07-31
Did anyone had a chance to test with the 2.7.3-SNAPSHOT build I released to see if the performance issue has been addressed?  If it has then let me know and I'll release the final 2.7.3 build.