@jbax 2018-08-01
Thanks @MaxGekk I've fixed the error and also made the parser run faster than before when processing fields that were not selected in general. 

Can you please retest with the latest SNAPSHOT build and let me know how it goes?