@MaxGekk 2018-08-02
I opened the separate PR for switching on **2.7.3**. Please, take a look at #21969